Added Layers for Obstacle Detection
1. Obstacle avoidance is now layer based. Different layers can be set for individual agent and block. Agents detect obstacle based on it's own layer detection list, which can be set both in editer and runtime.
2. The overall layers setting is now stored in a scriptableobject AStarSettings.

+++ Assets/AStar/Scripts/MoveAgent.cs
@@ -12,11 +12,9 @@ namespace AStar
     public enum AgentState
     {
         IDLE,       //Initial state
-        //PAUSED,     //Step by step movement, waiting for resume command
         TARGETING,  //Finding destination tile
         TARGETED,   //Destination tile set and fund
         ROUTING,    //Finding path, waiting for A* algorithm to finish
-        //MOVING,     //Agent is moving
         ROUTED,     //A* finished
         ARRIVING,   //Arrived at destination. Sending arrived message at next frame
         ARRIVED,    //Arrived. Message sent
@@ -29,16 +27,20 @@ namespace AStar
         protected int m_id;
         protected List<AStarTile> m_waypointTiles;
 
+        [Header("Astar Tile Layer Mask")]
+        [SerializeField] protected LayerMask m_tileLayerMask;
         [Header("Optional")]
         [SerializeField] protected PathOverlay m_pathOverlay = null;
         [Header("Transform to move")]
         [SerializeField] protected Transform m_moveTrans = null;
         [Header("Step by step?")]
-        [SerializeField] protected bool m_descreteMovement = false;
+        [SerializeField] public bool descreteMovement = false;
         [Header("Path Generator")]
         [SerializeField] protected PathGenerator m_pathGenerator = null;
         [Header("Obstacle Avoidance Mode")]
-        [SerializeField] public OAMode m_oaMode = OAMode.ACTIVE;
+        [SerializeField] public OAMode oaMode = OAMode.ACTIVE;
+        [Header("Obstacle Settings")]
+        [SerializeField] protected AgentObstacleSetting m_obstacleSetting = null;
 
         private AStarPath m_tempPath;
         private AStarTile m_targetTile;
@@ -51,20 +53,12 @@ namespace AStar
         private Vector3 m_lastLocation;
         private int m_routeCount;
         private bool m_astarFlag;
+        [SerializeField]private AStarLayer m_astarLayer;
+        private List<AStarLayer> m_obstacleLayers;
+
 
         public UnityEvent OnArrival { get { return m_onArrival; } }
         public AgentState AgentState { get; private set; }
-        public bool IsDiscrete
-        {
-            get
-            {
-                return m_descreteMovement;
-            }
-            set
-            {
-                m_descreteMovement = value;
-            }
-        }
         public bool IsOverlayActive
         {
             get
@@ -83,7 +77,7 @@ namespace AStar
                 if (AgentState == AgentState.IDLE)
                 {
                     RaycastHit r_hit;
-                    if (Physics.Raycast(new Vector3(m_moveTrans.position.x, 1f, m_moveTrans.position.z), Vector3.down, out r_hit, 2f, LayerMask.GetMask("AstarTile")))
+                    if (Physics.Raycast(new Vector3(m_moveTrans.position.x, 1f, m_moveTrans.position.z), Vector3.down, out r_hit, 2f, m_tileLayerMask))
                     {
                         m_currentTile = r_hit.transform.GetComponent<AStarTile>();
                     }
@@ -115,11 +109,9 @@ namespace AStar
         [System.Serializable]
         public enum OAMode
         {
-            PASSIVE,
             ACTIVE,
+            PASSIVE,
         }
-
-
 
         public static T1 AddAgent<T1,T2>(GameObject g_obj) where T1:MoveAgent where T2:PathGenerator
         {
@@ -138,7 +130,7 @@ namespace AStar
         {
             m_agents.Remove(agent.m_id);
             agent.m_pathOverlay.ActiveOverlay(false);
-            agent.CurrentTile.TileType = TileType.BLANK;
+            agent.CurrentTile.Layer = agent.CurrentTile.InitialLayer;
             if (remove_gameObject)
                 Destroy(agent.gameObject);
             else
@@ -155,7 +147,7 @@ namespace AStar
                     MoveAgent ma = agent;
                     keys.Add(agent.m_id);
                     agent.m_pathOverlay.ActiveOverlay(false);
-                    agent.CurrentTile.TileType = TileType.BLANK;
+                    agent.CurrentTile.Layer = agent.CurrentTile.InitialLayer;
                     if (remove_gameObject)
                         Destroy(ma.gameObject);
                     else
@@ -200,7 +192,7 @@ namespace AStar
                 if (m_astarThread != null)
                     m_astarThread.Abort();
 
-                s_tile.TileType = TileType.AGENT;
+                s_tile.Layer = m_astarLayer;
                 s_tile.Agent = this;
                 m_astarTime = 0;
                 m_astarThread = new Thread(TargetRoute);
@@ -232,8 +224,8 @@ namespace AStar
             if (Mathf.Abs(m_instantSpeed) <= float.Epsilon && m_tempPath.Successful)
                 return;
 
-            m_descreteMovement = true;
-            if (m_oaMode == OAMode.PASSIVE)
+            descreteMovement = true;
+            if (oaMode == OAMode.PASSIVE)
             {
                 m_waypointTiles = new List<AStarTile>(m_tempPath.PathTiles);
 
@@ -248,11 +240,22 @@ namespace AStar
                 if (m_waypointTiles.Count > 2)
                     m_waypointTiles.RemoveRange(2, m_waypointTiles.Count - 2);
             }
-            else if (m_oaMode == OAMode.ACTIVE)
+            else if (oaMode == OAMode.ACTIVE)
             {
                 if (m_waypointTiles != null && m_waypointTiles.Count >= 2)
                     m_astarFlag = true;
             }
+        }
+
+        //return true if the tile is within an obstacle astar layer
+        public bool CheckObstacle(AStarTile tile)
+        {
+            for (int i = 0; i < m_obstacleLayers.Count; i++)
+            {
+                if (tile.Layer == m_obstacleLayers[i] && tile.Agent != this)
+                    return true;
+            }
+            return false;
         }
 
         private void Awake()
@@ -262,6 +265,13 @@ namespace AStar
             m_agents[m_agentCount] = this;
             m_agentCount++;
             m_onArrival = new UnityEvent();
+            m_astarLayer = AStarManager.SetAStarLayer(m_obstacleSetting.astarLayer.layerID);
+            int count = m_obstacleSetting.obstacleAstarLayers.Count;
+            m_obstacleLayers = new List<AStarLayer>();
+            for (int i = 0; i < count; i++)
+            {
+                m_obstacleLayers.Add(AStarManager.SetAStarLayer(m_obstacleSetting.obstacleAstarLayers[i].layerID));
+            }
         }
 
         //movement state machine
@@ -303,14 +313,14 @@ namespace AStar
                     }
                     
                     //Passive mode. Start finding path only if the original path is blocked
-                    if(m_oaMode == OAMode.PASSIVE)
+                    if(oaMode == OAMode.PASSIVE)
                     {
                         m_astarFlag = false;
                         int count = m_waypointTiles.Count;
                         int i = 1;
                         while (i < count)
                         {
-                            if (m_waypointTiles[i].TileType == TileType.BLOCK || (m_waypointTiles[i].TileType == TileType.AGENT && m_waypointTiles[i].Agent != this))
+                            if (CheckObstacle(m_waypointTiles[i]))
                             {
                                 m_astarFlag = true;
                                 break;
@@ -320,7 +330,7 @@ namespace AStar
                     }
 
                     //Active mode. Start path finding when the distance to the next tile is small enough
-                    if (!m_descreteMovement && m_oaMode == OAMode.ACTIVE)
+                    if (!descreteMovement && oaMode == OAMode.ACTIVE)
                     {
                         if (dist <= m_instantSpeed * m_astarTime + 0.01f && m_routeCount == 0)
                         {
@@ -340,13 +350,13 @@ namespace AStar
                         if (tile.Agent == this)
                         {
                             tile.Agent = null;
-                            tile.TileType = TileType.BLANK;
+                            tile.Layer = tile.InitialLayer;
                         }
                     }
                     if (m_waypointTiles.Count > 1 && m_waypointTiles[1].Agent == null)
                     {
                         m_waypointTiles[1].Agent = this;
-                        m_waypointTiles[1].TileType = TileType.AGENT;
+                        m_waypointTiles[1].Layer = m_astarLayer;
                     }
                 }
 
@@ -390,7 +400,7 @@ namespace AStar
             m_tempPath = m_pathGenerator.GeneratePath(Grid, CurrentTile, m_targetTile, this);
             m_overlayUpdated = false;
 
-            if (!m_descreteMovement && m_tempPath.Successful)
+            if (!descreteMovement && m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>(m_tempPath.PathTiles);
                 int index = -1;
@@ -401,7 +411,7 @@ namespace AStar
 
                 m_waypointTiles.RemoveRange(0, index);
             }
-            else if (m_descreteMovement && m_tempPath.Successful)
+            else if (descreteMovement && m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>();
             }
@@ -409,13 +419,13 @@ namespace AStar
             {
                 if (m_waypointTiles != null)
                 {
-                    m_waypointTiles[0].TileType = TileType.AGENT;
+                    m_waypointTiles[0].Layer = m_astarLayer;
                     m_waypointTiles[0].Agent = this;
-                    m_waypointTiles[1].TileType = TileType.BLANK;
+                    m_waypointTiles[1].Layer = m_waypointTiles[1].InitialLayer;
                 }
                 m_waypointTiles = new List<AStarTile>();
             }
-            m_astarFlag = !m_descreteMovement;
+            m_astarFlag = !descreteMovement;
             //m_astarTime = Time.unscaledTime - startTime;
             if (m_astarTime < m_deltaTime && m_tempPath.Successful)
                 m_astarFlag = false;
@@ -424,10 +434,9 @@ namespace AStar
 
         private void AstarWorker1()
         {
-            //Debug.Log("1");
             m_tempPath = m_pathGenerator.GeneratePath(Grid, CurrentTile, m_targetTile,this);
             m_overlayUpdated = false;
-            if (!m_descreteMovement && m_tempPath.Successful)
+            if(m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>(m_tempPath.PathTiles);
 
@@ -439,27 +448,15 @@ namespace AStar
                 } while (index < m_tempPath.PathTiles.Count && m_tempPath.PathTiles[index] != m_currentTile);
 
                 m_waypointTiles.RemoveRange(0, index);
-            }
-            else if (m_descreteMovement && m_tempPath.Successful)
-            {
-                m_waypointTiles = new List<AStarTile>(m_tempPath.PathTiles);
 
-                //remove excessive tiles in the path caused by delay
-                int index = -1;
-                do
-                {
-                    index++;
-                } while (index < m_tempPath.PathTiles.Count && m_tempPath.PathTiles[index] != m_currentTile);
-
-                m_waypointTiles.RemoveRange(0, index);
-                if (m_waypointTiles.Count > 2)
+                if (m_waypointTiles.Count > 2 && descreteMovement)
                     m_waypointTiles.RemoveRange(2, m_waypointTiles.Count - 2);
             }
             else if (!m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>();
             }
-            m_astarFlag = !m_descreteMovement;
+            m_astarFlag = !descreteMovement;
             if (m_astarTime < m_deltaTime && m_tempPath.Successful)
                 m_astarFlag = false;
             AgentState = AgentState.ROUTED;
@@ -473,7 +470,7 @@ namespace AStar
             AgentState = AgentState.ROUTING;
             m_tempPath = m_pathGenerator.GeneratePath(Grid, CurrentTile, m_targetTile,this);
 
-            if (!m_descreteMovement && m_tempPath.Successful)
+            if (!descreteMovement && m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>(m_tempPath.PathTiles);
                 int index = -1;
@@ -483,7 +480,7 @@ namespace AStar
                 } while (index < m_tempPath.PathTiles.Count && m_tempPath.PathTiles[index] != m_currentTile);
                 m_waypointTiles.RemoveRange(0, index);
             }
-            else if (m_descreteMovement && m_tempPath.Successful)
+            else if (descreteMovement && m_tempPath.Successful)
             {
                 m_waypointTiles = new List<AStarTile>();
                 m_waypointTiles.Add(m_tempPath.PathTiles[0]);
@@ -495,9 +492,9 @@ namespace AStar
 
         private void OnDestroy()
         {
-            CurrentTile.TileType = TileType.BLANK;
+            CurrentTile.Layer = CurrentTile.InitialLayer;
             if (AgentState != AgentState.IDLE && NextTile && NextTile.Agent == this)
-                NextTile.TileType = TileType.BLANK;
+                NextTile.Layer = NextTile.InitialLayer;
         }
     }
 }
